[ANDROSDK-1386] Refactor packages
