+++ core/src/main/java/org/hisp/dhis/android/core/analytics/aggregated/internal/AnalyticsException.kt
@@ -26,36 +26,8 @@
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.analytics.aggregated
+package org.hisp.dhis.android.core.analytics.aggregated.internal
 
-import io.reactivex.Single
-import javax.inject.Inject
-import org.hisp.dhis.android.core.analytics.aggregated.service.AnalyticsService
-
-internal class AnalyticsRepositoryImpl @Inject constructor(
-    private val params: AnalyticsRepositoryParams,
-    private val analyticsService: AnalyticsService
-) : AnalyticsRepository {
-
-    override fun withDimension(dimensionItem: DimensionItem): AnalyticsRepositoryImpl {
-        return updateParams { params -> params.copy(dimensions = params.dimensions + dimensionItem) }
-    }
-
-    override fun withFilter(dimensionItem: DimensionItem): AnalyticsRepositoryImpl {
-        return updateParams { params -> params.copy(filters = params.filters + dimensionItem) }
-    }
-
-    override fun evaluate(): Single<DimensionalResponse> {
-        return Single.fromCallable { blockingEvaluate() }
-    }
-
-    override fun blockingEvaluate(): DimensionalResponse {
-        return analyticsService.evaluate(params)
-    }
-
-    private fun updateParams(
-        func: (params: AnalyticsRepositoryParams) -> AnalyticsRepositoryParams
-    ): AnalyticsRepositoryImpl {
-        return AnalyticsRepositoryImpl(func(params), analyticsService)
-    }
+sealed class AnalyticsException(message: String) : Throwable(message) {
+    class InvalidArguments(message: String) : AnalyticsException(message)
 }
