+++ core/src/main/java/org/hisp/dhis/android/core/analytics/aggregated/internal/AnalyticsRepositoryImpl.kt
@@ -26,15 +26,38 @@
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.analytics.aggregated.service.evaluator
+package org.hisp.dhis.android.core.analytics.aggregated.internal
 
-import org.hisp.dhis.android.core.analytics.aggregated.MetadataItem
-import org.hisp.dhis.android.core.analytics.aggregated.service.AnalyticsServiceEvaluationItem
+import io.reactivex.Single
+import org.hisp.dhis.android.core.analytics.aggregated.AnalyticsRepository
+import org.hisp.dhis.android.core.analytics.aggregated.DimensionItem
+import org.hisp.dhis.android.core.analytics.aggregated.DimensionalResponse
+import javax.inject.Inject
 
-internal interface AnalyticsEvaluator {
+internal class AnalyticsRepositoryImpl @Inject constructor(
+    private val params: AnalyticsRepositoryParams,
+    private val analyticsService: AnalyticsService
+) : AnalyticsRepository {
 
-    fun evaluate(
-        evaluationItem: AnalyticsServiceEvaluationItem,
-        metadata: Map<String, MetadataItem>
-    ): String?
+    override fun withDimension(dimensionItem: DimensionItem): AnalyticsRepositoryImpl {
+        return updateParams { params -> params.copy(dimensions = params.dimensions + dimensionItem) }
+    }
+
+    override fun withFilter(dimensionItem: DimensionItem): AnalyticsRepositoryImpl {
+        return updateParams { params -> params.copy(filters = params.filters + dimensionItem) }
+    }
+
+    override fun evaluate(): Single<DimensionalResponse> {
+        return Single.fromCallable { blockingEvaluate() }
+    }
+
+    override fun blockingEvaluate(): DimensionalResponse {
+        return analyticsService.evaluate(params)
+    }
+
+    private fun updateParams(
+        func: (params: AnalyticsRepositoryParams) -> AnalyticsRepositoryParams
+    ): AnalyticsRepositoryImpl {
+        return AnalyticsRepositoryImpl(func(params), analyticsService)
+    }
 }
