+++ core/src/main/java/org/hisp/dhis/android/core/analytics/aggregated/internal/AnalyticsRepositoryParams.kt
@@ -26,8 +26,11 @@
  *  SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.analytics.aggregated.service
+package org.hisp.dhis.android.core.analytics.aggregated.internal
 
-sealed class AnalyticsException(message: String) : Throwable(message) {
-    class InvalidArguments(message: String) : AnalyticsException(message)
-}
+import org.hisp.dhis.android.core.analytics.aggregated.DimensionItem
+
+internal data class AnalyticsRepositoryParams(
+    val dimensions: List<DimensionItem>,
+    val filters: List<DimensionItem>
+)
